requirements.txt
fastapi
uvicorn
pydantic
faiss-cpu
sentence-transformers
transformers
torch
numpy
pdfplumber
python-docx
pandas
python-multipart
